requirements.txt
arch==6.2.0
matplotlib==3.8.2
numpy==1.23.5
pandas==2.2.0
PyYAML==6.0.1
scipy==1.12.0
tqdm==4.66.1
